instagram-web-downloader/requirements.txt
Flask==3.0.0
instaloader>=4.9.6
yt-dlp>=2023.12.30
Pillow>=9.5.0
requests>=2.28.0
Werkzeug>=3.0.0
gunicorn>=21.2.0
